Backing-out unittest.TestCase.
